substitution resolver module decoupling, added a few tests for breaking cycle in self-referential substitutions
