+++ tests/acceptance/includes/data/1.conf
@@ -1,0 +1,1 @@
+house.location: ${?house.city} "Nice street 42"
