+++ tests/acceptance/includes/data/2.conf
@@ -1,0 +1,3 @@
+house {
+  city: "서울"
+  }
