+++ tests/acceptance/includes/data/main.conf
@@ -1,0 +1,4 @@
+{
+  include "1.conf"
+  include "2.conf"
+}
