+++ tests/ut/resolver/test_merge.py
